backend/requirements.txt
Flask
requests
imap-tools
python-dotenv
pymongo

